+++ Sonic Mania/Objects/UFO_Springboard.c
@@ -61,7 +61,7 @@ void UFO_Springboard_LateUpdate(void)
 
     entity->depth = mat->values[2][1] * (y >> 16) + mat->values[2][2] * (z >> 16) + mat->values[2][0] * (x >> 16) + mat->values[2][3];
     if (entity->depth >= 0x4000) {
-        int depth = ((mat->values[0][3] << 8) + (mat->values[0][2] * (z >> 8) & 0xFFFFFF00) + (mat->values[0][0] * (x >> 8) & 0xFFFFFF00)
+        int depth = (int)((mat->values[0][3] << 8) + (mat->values[0][2] * (z >> 8) & 0xFFFFFF00) + (mat->values[0][0] * (x >> 8) & 0xFFFFFF00)
                      + (mat->values[0][1] * (entity->height >> 8) & 0xFFFFFF00))
                     / entity->depth;
         entity->visible = abs(depth) < 256;
